# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dogfood 2026 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dogfood_session"
      }
    }
  },
  "paths": {
...
    "/api/judge/records": {
      "get": {
        "summary": "Get a signed, aggregate-only judge participation record (assigned/completed counts, no scores) — ?eventId=&judge=",
        "tags": [
          "signing"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "format": {
                      "type": "undefined",
                      "enum": [
                        null
                      ]
                    },
                    "version": {
                      "type": "undefined",
                      "enum": [
                        null
                      ]
                    },
                    "type": {
                      "type": "undefined",
                      "enum": [
                        null
                      ]
                    },
                    "alg": {
                      "type": "undefined",
                      "enum": [
                        null
                      ]
                    },
                    "kid": {
                      "type": "string"
                    },
                    "signed_at": {
                      "type": "string"
                    },
                    "payload": {
                      "type": "object",
                      "properties": {
                        "eventId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "judgeId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "judgeName": {
                          "type": "string",
                          "nullable": true
                        },
                        "assignedCount": {
                          "type": "integer"
                        },
                        "completedCount": {
                          "type": "integer"
                        },
                        "firstScoredAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "lastScoredAt": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "eventId",
                        "judgeId",
                        "judgeName",
                        "assignedCount",
                        "completedCount",
                        "firstScoredAt",
                        "lastScoredAt"
                      ]
                    },
                    "signature": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "format",
                    "version",
                    "type",
                    "alg",
                    "kid",
                    "signed_at",
                    "payload",
                    "signature"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired session"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  }
}